requirements.txt
Flask==3.0.2
requests==2.25.1
beautifulsoup4==4.9.3
python-dotenv==0.19.0
openai==0.10.2
